Rename DirectClient to Peer because that's what it is
The name direct client is confusing to some people that it make them think it's a client as in the nomral sense of client server not that it's a direct client to the relay which makes it a peer

A peer (unlike a client) can also receive any message that is sent to it by other peers.

This allows us to build cool things with this:

- RpcClient when u wanna use the peer to only make requests (and receive responses)
- Peer by itself then provide you the functionality to also serve requests to peer by implementing the proper handler

TBD: A Router handler. which is a special handler that can be used by developer to quickly build a router to handle specific commands. This should work as follows:

- When a "Request" message is received, the router matches the command to the proper handle function, passing only the proper context, and the request body.
- When the handler return, the router will create the proper envelope from the content and send it back to caller.