This introduces 2 different cache policy for twin for rmb client
the default behavior compatible with previous behaviour. Keeps twin
information in memory forever.

we now also introduce the option peer.WithTwinCache which can customize
how long the twin cache is kept in cache before refreshing

The rest of the change is just about using the proper way to customize
the client using options not a long argument list