fix node status ordering by using different gorm query for counting
according to [gorm docs](https://gorm.io/docs/method_chaining.html#Reusability-and-Safety) the retruned `*gorm.DB` instance of a finisher method is unsafe to use.

the count query was happening before the ordering and then the same `*gorm.DB` instance is reused. so creating new instance of the query chain before executing the finisher method `.Count()` should leave the `q` instance with all the chain methods for filtering a safe to reuse.

- rearrange the counting code after the ordering part
- use a different query instance for the finisher method count.